+++ SocialPlatformProjectWebApi/SocialPlatformProjectWebApi/Repository/ReplyRepository.cs
@@ -43,36 +43,23 @@ namespace SocialPlatformProjectWebApi.Repository
             return reply;
         }
 
-        public async Task<Reply> AddReply(Reply reply)
+        public async Task<bool> AddReply(Reply template)
         {
-            DateTime date = DateTime.Now;
-
-            //var incomingObject = reply;
-            //_dbContext.Replies.AddAsync(incomingObject);
-            //_dbContext.SaveChangesAsync();
-
-            var template = new Reply
-            {
-                Text = reply.Text,
-                CreatedDate = date,
-                CategoryThreadId = reply.CategoryThreadId,
-                UserIdSub = reply.UserIdSub,
-            };
-
             await _dbContext.AddAsync(template);
-            await _dbContext.SaveChangesAsync();
-
-            return template;
+            return (await _dbContext.SaveChangesAsync() > 0);
         }
 
-        public async Task<Reply> DeleteReply(int id)
+        public async Task<bool> DeleteReply(int id)
         {
             var deleteReply = await _dbContext.Replies.SingleAsync(x => x.Id == id);
+
+            if(deleteReply == null)
+            {
+                return false;
+            }
+
             _dbContext.Replies.Remove(deleteReply);
-
-            await _dbContext.SaveChangesAsync();
-
-            return deleteReply;
+            return (await _dbContext.SaveChangesAsync() > 0);
         }
 
         public async Task<Reply> EditReplyText(int id, string text)
